Removes dead components
TL;DR
-----

Deletes artifacts that aren't needed with the current lab setup

Details
-------

As I documented the lab setup I realized that the best approach
was to build on the Tanzu Governance Lab rather than to replicate
steps in both places. That decision left a few artifacts in this
repository that weren't needed any longer. This change removes the
redundant artifacts.